requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
cffi==1.12.3
chardet==3.0.4
coverage==4.5.3
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Fabric3==1.14.post1
importlib-metadata==0.18
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.1.0
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
packaging==19.0
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
pluggy==0.12.0
prometheus-client==0.7.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.4.2
PyNaCl==1.3.0
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==5.0.1
pytest-cov==2.7.1
python-dateutil==2.8.0
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.5.1
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
zipp==0.5.2
tabulate